requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
libusb1==3.1.0
ndeflib==0.3.3
nfcpy==1.0.4
psycopg2-binary==2.9.9
pyDes==2.0.1
PyJWT==2.9.0
pyserial==3.5
sqlparse==0.5.1
tzdata==2024.1
